new clippy linds and faceoff example

+++ src/util/mod.rs
@@ -13,7 +13,7 @@ impl Util {
     pub fn calculate_percentage(number: usize, total: usize) -> f32 {
         match total {
             0 => 0_f32,
-            _ => ((number as f32 * 100.0) / total as f32) as f32,
+            _ => (number as f32 * 100.0) / total as f32,
         }
     }
 }
